Remove unused activity types and integrate EngagementIndicator component in TrackList
- Deleted the `activity.ts` and its TypeScript definition file as they were no longer needed.
- Added a new `EngagementIndicator` component to visually represent track engagement levels based on plays and downloads.
- Updated the `TrackList` component to utilize the new `EngagementIndicator`, improving the display of engagement metrics.
