--- packages/core-storage/src/types/activity.ts
@@ -1,1 +1,0 @@
-
